Common: Merge CRC32.h into Hash.h
This makes it easier to find the relevant functions.
