+++ Source/Core/Common/CMakeLists.txt
@@ -25,8 +25,6 @@ add_library(common
   Config/Layer.cpp
   Config/Layer.h
   CPUDetect.h
-  CRC32.cpp
-  CRC32.h
   Crypto/AES.cpp
   Crypto/AES.h
   Crypto/bn.cpp
